Fixed Mac build

+++ include/IClientDriver.h
@@ -18,7 +18,7 @@ namespace SL {
 			virtual void onReceive_ImageDif(const Image& img, int monitor_id) = 0;
 			virtual void onReceive_MouseImage(const Image& img) = 0;
 			virtual void onReceive_MousePos(const Point* pos) = 0;
-			virtual void onReceive_ClipboardText(const char* data, size_t length) = 0;
+			virtual void onReceive_ClipboardText(const char* data, unsigned int length) = 0;
 
 		};
